move to util so it can be accessed later when we write eval.py
