+++ train.py
@@ -8,6 +8,7 @@ from keras import backend as k
 
 from constants import *
 from hourglass import HourglassNet
+from util import *
 
 # TODO
 # Add command line parameter for learning rate?
@@ -81,26 +82,6 @@ def process_args():
         find_resume_json_weights(args)
 
     return args
-
-def find_resume_json_weights(args):
-    enclosing_dir = os.path.join(args.model_save, args.resume_subdir)
-
-    files = os.listdir(enclosing_dir)
-
-    model_jsons         = [f for f in files if (".json" in f and "hpe_hourglass_stacks" in f)]
-    model_saved_weights = [f for f in files if (".hdf5" in f and "hpe_epoch{epoch:02d}".format(epoch=args.resume_epoch) in f)]
-    
-    assert len(model_jsons) > 0
-    assert len(model_saved_weights) > 0
-
-    args.resume_json = os.path.join(enclosing_dir, model_jsons[0])
-    args.resume_weights = os.path.join(enclosing_dir, model_saved_weights[0])
-
-    print('Found model json: {}'.format(args.resume_json))
-    print('Found model weights for epoch {epoch:02d}: {weight_file_name}'.format(epoch=args.resume_epoch, weight_file_name=args.resume_weights))
-
-    assert os.path.exists(args.resume_json)
-    assert os.path.exists(args.resume_weights)
 
 if __name__ == "__main__":
     args = process_args()
